+++ RELEASE-NOTES.md
@@ -29,6 +29,7 @@ It also brings some dreadfully awaited fixes, so be sure to go through the chang
 - Fixed mathematical formulation in `MvStudentT` random method. (see [#4359](https://github.com/pymc-devs/pymc3/pull/4359))
 - Fix issue in `logp` method of `HyperGeometric`. It now returns `-inf` for invalid parameters (see [4367](https://github.com/pymc-devs/pymc3/pull/4367))
 - Fixed `MatrixNormal` random method to work with parameters as random variables. (see [#4368](https://github.com/pymc-devs/pymc3/pull/4368))
+- Fix issue in `logcdf` methods of `Uniform`, `HalfNormal`, `Gamma`  and `InverseGamma`. These functions now return correct values when evaluated with invalid parameters or values (see [4393](https://github.com/pymc-devs/pymc3/pull/4393))
 
 ## PyMC3 3.10.0 (7 December 2020)